新Firebase Admin SDKに対応: append系
