+++ contestmanager/append-winners.js
@@ -1,8 +1,8 @@
 /**
- * append-kaikin.js
+ * append-winners.js
  *
  * シーズンが終わるごとに実行する。
- * MySQL の kaikin テーブルに存在する待ちレコードに対して、実際にストアポイントを追加する。
+ * MySQL の winners テーブルに存在する待ちレコードに対して、実際にストアポイントを追加する。
  */
 
 var async = require('async');
@@ -11,16 +11,7 @@ var exec = require('child_process').exec;
 
 var Config = require('./config.js');
 
-var Firebase = require('firebase');
-var contestRef = new Firebase('https://' + Config.CONTESTAPP + '.firebaseio.com/');
-
-// Create admin user
-var FirebaseTokenGenerator = require('firebase-token-generator');
-var tokenGenerator = new FirebaseTokenGenerator(Config.CONTESTAPP_SECRET);
-var token = tokenGenerator.createToken(
-    { uid: '1', some: 'arbitrary', data: 'here' },
-    { admin: true, debug: true }
-);
+var contestRef = require('./contestref.js').ref;
 
 var Contests, Events;
 
@@ -41,10 +32,10 @@ var connectionStore = mysql.createConnection({
 });
 connectionStore.connect();
 
-// MySQL lottery_log テーブルの待ちレコードを対象にして、抽選ポイント加算
+// MySQL winners テーブルの待ちレコードを対象にして、抽選ポイント加算
 var appendPoints = function() {
     // ポイント加算履歴を検索して、未加算のものを対象とする。
-    connection.query('SELECT id, user_id, season, event_id, customer_type, customer_id, point FROM kaikin WHERE appended_at IS NULL', function(error, results, fields) {
+    connection.query('SELECT id, user_id, season, event_id, place, customer_type, customer_id, point FROM winners WHERE appended_at IS NULL', function(error, results, fields) {
         if (error) {
             console.error(error);
         } else {
@@ -56,11 +47,11 @@ var appendPoints = function() {
                 if (!(result.user_id in Ready)) {
                     Ready[result.user_id] = result;
                     Ready[result.user_id]['events'] = [result.event_id];
-                    Ready[result.user_id]['events_name'] = [Events[result.event_id].name];
+                    Ready[result.user_id]['events_name'] = [Events[result.event_id].name + '_' + result.place + '位_' + result.point + 'ポイント'];
                     Ready[result.user_id]['point_total'] = result.point;
                 } else {
                     Ready[result.user_id]['events'].push(result.event_id);
-                    Ready[result.user_id]['events_name'].push(Events[result.event_id].name);
+                    Ready[result.user_id]['events_name'].push(Events[result.event_id].name + '_' + result.place + '位_' + result.point + 'ポイント');
                     Ready[result.user_id]['point_total'] += result.point;
                 }
                 next();
@@ -102,7 +93,7 @@ var appendPoints = function() {
                                     process.exit(1);
                                 } else {
                                     console.log('Succeeded updating point!');
-                                    connection.query('UPDATE kaikin SET appended_at = NOW() WHERE appended_at IS NULL AND user_id = ?', [
+                                    connection.query('UPDATE winners SET appended_at = NOW() WHERE appended_at IS NULL AND user_id = ?', [
                                         r.user_id
                                     ], function(error, results, fields) {
                                         if (error) {
@@ -116,7 +107,7 @@ var appendPoints = function() {
                                             console.log(r.season);
                                             console.log(r.events_name.join('-'));
                                             console.log(r.point_total);
-                                            var command = '/usr/bin/php ' + __dirname + '/send-kaikin.php'
+                                            var command = '/usr/bin/php ' + __dirname + '/send-winners.php'
                                                         + ' "' + email + '"'
                                                         + ' "' + name + '"'
                                                         + ' "' + r.season + '"'
@@ -158,20 +149,11 @@ var appendPoints = function() {
 };
 
 var getEventsInfo = function() {
-    // admin 権限でログインしてから操作する
-    contestRef.authWithCustomToken(token, function(error, authData) {
-        if (error) {
-            console.error('Authentication Failed!', error);
-        } else {
-            //console.log('Authenticated successfully with payload:', authData);
-
             contestRef.child('events').once('value', function(snapEvents) {
                 Events = snapEvents.val();
 
                 appendPoints();
             });
-        }
-    });
 };
 
 var main = function() {
